+++ database/migrations/2025_02_04_024645_create_tagihanxis_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('tagihanxis', function (Blueprint $table) {
             $table->id();
             $table->string('judul_XI');
-            $table->float('harga_XI');
+            $table->integer('harga_XI');
             $table->string('kelas_XI');
             $table->timestamps();
         });
