+++ database/migrations/2025_02_05_013839_create_tagihanxiis_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('tagihanxiis', function (Blueprint $table) {
             $table->id();
             $table->string('judul_XII');
-            $table->float('harga_XII');
+            $table->integer('harga_XII');
             $table->string('kelas_XII');
             $table->timestamps();
         });
